backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
Django==5.2.8
django-bootstrap5==25.2
django-braces==1.17.0
django-cors-headers==4.9.0
django-esewa==1.0.9
django-model-utils==5.0.0
django-star-ratings==0.9.2
django-widget-tweaks==1.4.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
msgpack==1.1.2
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.3
sqlparse==0.5.3
swapper==1.4.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
